requirements.txt
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.10.0
djoser==2.1.0
drf-yasg==1.20.0
environs==9.3.5
gunicorn==20.1.0
psycopg2==2.9.2
